feat: added info modal for search + refactor modals structure
