+++ client/src/app/components/header/Header.scss
@@ -17,6 +17,10 @@ header {
       background: var(--shimmer-navbar-bg);
   }
 
+  &.full-height {
+    height: 100vh;
+  }
+
   a {
     outline: none;
     text-decoration: none;
