+++ client/src/app/components/stardust/MilestonePayload.tsx
@@ -11,7 +11,7 @@ import { StardustTangleCacheService } from "../../../services/stardust/stardustT
 import AsyncComponent from "../../components/AsyncComponent";
 import Modal from "../../components/Modal";
 import ReceiptPayload from "../stardust/ReceiptPayload";
-import milestoneMessage from "./../../../assets/modals/message/milestone-payload.json";
+import milestoneMessage from "./../../../assets/modals/block/milestone-payload.json";
 import "./MilestonePayload.scss";
 import { MilestonePayloadProps } from "./MilestonePayloadProps";
 import { MilestonePayloadState } from "./MilestonePayloadState";
